Perform basic style reset
Most modern browsers are more consistent in how they render a website,
making more verbose resets (e.g., normalize.css) still valid, but less
necessary.

This approach uses the universal selector to remove any margin and
padding from all elements and defaults the box-sizing property to
border-box so that the height and width of an element includes the
content, padding, and border.

https://www.w3schools.com/cssref/css3_pr_box-sizing.asp