Set header background size and position
Setting the background size to `cover` causes the background image to
fill the entirety of its container, even if it has to stretch the image
or cut off a little bit of one of the edges.

https://www.w3schools.com/cssref/css3_pr_background-size.asp

Setting the background position to `top` (i.e., `center top`) will
ensure that the top of the image is anchored as the image is resized.