Add visual animation to header callout
For best performance, only ever animate the `opacity` and/or the
`transform` properties; browsers are optimized for performance on these
two properties.

Using an animation timing function allows you to control the speed of
the animation from beginning to end. There are multiple built-in timing
functions, detailed in the link below. Here we use `ease-out` which
causes a slow end to the animation (starts fast).

https://www.w3schools.com/cssref/css3_pr_animation-timing-function.asp