Add header call to action button
This uses the transition effect to animate clicking the button. To
create a transition effect, you must specify the css property you want
to add the effect to and the duration of the effect (default: 0)

Adding a larger high blur box shadow on hover that changes to a
smaller low blur box shadow on active (click) is a nice effect for an
animated button.