Switch from px to rem
Rem units are relative to the root font size. In our case, we set the
root font size to 62.5%, which, on most browsers with a default font
setting of 16px, is equal to 10px. This trick is done for easier math in
multiples of 10.

There is some discussion on this particular technique and on when and
how to use rem units, particularly versus em units, which are relative
to the current element's font size or width, depending on which property
they are used.

Some additional reading is required to better understand the rem/em, but
for the sake of continuing with this project, following the instructor's
current lead of using rem for all things.