Add the about section
This introduces the concept of utility classes; simple, generic classes
meant to be applied to a wide range of elements. e.g, consistent margins
across different elements, center text in different elements, etc.

This also experiements with the `background-clip` property, setting it
to text allows the background to come through as text, in our case a
linear gradient.

We use the :not selector to negate the selector, e.g.,
`:not(:last-child)` to apply styles to everything _but_ the last
child.