Add rotating card component
The backface-visibility property defines whether or not the back face of
an element should be visible when facing the user. The back face of an
element is a mirror image of the front face being displayed. This
property is useful when an element is rotated and lets you choose
whether the user should see the back face or not.

When absolute positioning an element, the relative positioned parent
will lose its height if not specified. This is similar to how a parent
element when its children are floated but it is not solved with a
clearfix.