Style front side of card
The box-decoration-break property specifies how the background, padding,
border, border-image, box-shadow, margin, and clip-path of an element
are applied when the box for the element is fragmented. In this case we
chose clone to that each fragment is rendered independently with the
specified border, padding, and margin wrapping each fragment.

https://developer.mozilla.org/en-US/docs/Web/CSS/box-decoration-break

The background-blend-mode property defines the blending mode of each
background layer. In this case we chose screen.

https://www.w3schools.com/cssref/pr_background-blend-mode.asp

The size of the image extends beyond the containing element's applied
border-radius. We use the `overflow: hidden` declaration to hide the
portion of the image that extends beyond the containing element, thereby
allowing the border-radius to display.