Add background video to stories section
The CSS object-fit property is used to specify how an img or video
should be resized to fit its container. Here we chose cover, which will
cut off the sides of the video, preserving the aspect ration, and also
fill the space.

Always specify multiple video sources as different browsers use
different formats. Another common practice is to display text as an
additional option which will only display when the browser is unable to
play any of the source video formats (e.g., IE8).